test(ios): split adherence tests into their own file (SwiftLint file_length)
